lecture-surveys/pyproject.toml
[project]
name = "lecture-surveys"
version = "0.1.0"
description = "Surveys using FastHTML for the 2024 statistics lecture"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-fasthtml>=0.5.1",
    "qrcode>=7.4.2",
]

[tool.ruff]
line-length = 180

